前段改造，功能全部完善提交

+++ spider-business/src/main/java/com/spider/business/repostory/mapper/HaohuoCommodityMapper.java
@@ -7,11 +7,11 @@ import java.util.List;
 
 @MyBatisRepository
 public interface HaohuoCommodityMapper {
-    int deleteByPrimaryKey(Long id);
+    int deleteByPrimaryKey(Integer id);
 
     int insert(HaohuoCommodity record);
 
-    HaohuoCommodity selectByPrimaryKey(Long id);
+    HaohuoCommodity selectByPrimaryKey(Integer id);
 
     HaohuoCommodity selectByProductId(String productId);
